op code

+++ src/worker.js
@@ -9,30 +9,38 @@ const WorkerType = {
 };
 
 class Worker {
-  static spawn(type = WorkerType.WORKER_TYPE_CLUSTER) {
+  static spawn(type = WorkerType.WORKER_TYPE_CLUSTER, exitCB, errorCB, disconnectCB, messageCB, onlineCB) {
+    let worker;
+
     if (type === WorkerType.WORKER_TYPE_CLUSTER) {
       cluster.setupMaster({
         exec: path.resolve(__dirname, 'cluster-worker-task.js'),
         silent: true,
       });
-      const worker = cluster.fork();
-      return new Worker(worker, type);
+      worker = new Worker(cluster.fork(), type, exitCB, errorCB, disconnectCB, messageCB, onlineCB);
     } else if (type === WorkerType.WORKER_TYPE_THREAD) {
       const ThreadWorker = require('worker_threads').Worker;
-      const worker = new ThreadWorker(path.resolve(__dirname, 'thread-worker-task.js'));
-      return new Worker(worker, type);
+      worker = new Worker(new ThreadWorker(path.resolve(__dirname, 'thread-worker-task.js')), type, exitCB, errorCB, disconnectCB, messageCB, onlineCB);
     }
+    return worker;
   }
 
-  constructor(worker, type) {
+  constructor(worker, type, exitCB, errorCB, disconnectCB, messageCB, onlineCB) {
     this._worker = worker;
     this._type = type;
-    this._count = 0;
     this._runningTasks = [];
-  }
-
-  get count() {
-    return this._count;
+    this._exitCB = exitCB;
+    this._errorCB = errorCB;
+    this._disconnectCB = disconnectCB;
+    this._messageCB = messageCB;
+    this._onlineCB = onlineCB;
+    if (this._worker) {
+      if (this._exitCB) this._worker.on('exit', this._exitCB);
+      if (this._errorCB) this._worker.on('error', this._errorCB);
+      if (this._disconnectCB) this._worker.on('disconnect', this._disconnectCB);
+      if (this._messageCB) this._worker.on('message', this._messageCB);
+      if (this._onlineCB) this._worker.on('online', this._onlineCB);
+    }
   }
 
   get id() {
@@ -52,30 +60,37 @@ class Worker {
   }
 
   sendTask(task) {
-    if (this._type === WorkerType.WORKER_TYPE_CLUSTER) {
-      this._worker.send(task);
-    } else if (this._type === WorkerType.WORKER_TYPE_THREAD) {
-      this._worker.postMessage(task);
+    if (this._worker) {
+      if (this._type === WorkerType.WORKER_TYPE_CLUSTER) {
+        this._worker.send(task);
+      } else if (this._type === WorkerType.WORKER_TYPE_THREAD) {
+        this._worker.postMessage(task);
+      }
+      this._runningTasks.push(task);
     }
-    this._runningTasks.push(task);
-    this._count++;
   }
 
   receiveResult(result) {
-    this._count--;
     this._runningTasks = this._runningTasks.filter(task => task.msgID !== result.msgID);
   }
 
   terminate() {
-    if (this._type === WorkerType.WORKER_TYPE_CLUSTER) {
-      this._worker.kill();
-    } else if (this._type === WorkerType.WORKER_TYPE_THREAD) {
-      this._worker.terminate();
+    if (this._worker) {
+      if (this._exitCB) this._worker.removeListener('exit', this._exitCB);
+      if (this._errorCB) this._worker.removeListener('error', this._errorCB);
+      if (this._disconnectCB) this._worker.removeListener('disconnect', this._disconnectCB);
+      if (this._messageCB) this._worker.removeListener('message', this._messageCB);
+      if (this._onlineCB) this._worker.removeListener('online', this._onlineCB);
+      if (this._type === WorkerType.WORKER_TYPE_CLUSTER) {
+        this._worker.kill();
+      } else if (this._type === WorkerType.WORKER_TYPE_THREAD) {
+        this._worker.terminate();
+      }
+      const runningTasks = this._runningTasks;
+      this._runningTasks = [];
+      return runningTasks;
     }
-  }
-
-  on(...args) {
-    this._worker.on(...args);
+    return [];
   }
 }
